Clase 46, Entrega final, parte 1

+++ src/services/products.services.js
@@ -3,6 +3,12 @@ import ProductsDaoMongoDB from "../dao/mongodb/managers/products.dao.js"
 const prodDao = new ProductsDaoMongoDB();
 import { createProductsMocking } from "../utils/utils.js";
 
+// Agregado servicio de envio de correos
+import { sendMailEthereal } from "./email.services.js";
+
+// Módulo de métodos DAO de usuarios
+import UserDao from "../dao/mongodb/managers/user.dao.js";
+
 // Persistencia de archivos
 /* No se utilizará
 import { ProductManager } from "../dao/fs/ProductManager.js";
@@ -40,6 +46,26 @@ export const getByIdService = async (id) => {
     }
   }
 
+export const getByOwnerService = async (email) => {
+    try {
+      const response = await prodDao.getProductByOwner(email);
+      return response;
+    } catch (error) {
+      logger.error(error.message)
+      throw new Error(error.message)
+    }
+  }
+
+export const changeProdOwnerService = async (email) => {
+    try {
+      const response = await prodDao.changeProductOwner(email);
+      return response;
+    } catch (error) {
+      logger.error(error.message)
+      throw new Error(error.message)
+    }
+  }
+
 export const createService = async (obj) => {
     try {
       const response = await prodDao.createProduct(obj);
@@ -61,8 +87,16 @@ export const updateService = async (id, obj) => {
 
 export const deleteByIdService = async (id) => {
     try {
-      const response = await prodDao.deleteProduct(id);
-      return response;
+      const userDao = new UserDao();
+      const product = await getByIdService(id)
+      if ( product ) {
+        const response = await prodDao.deleteProduct(id);
+        if ( product.owner && product.owner !== 'admin' ) {
+          const user = await userDao.getByEmail(product.owner)
+          await sendMailEthereal( { destination: user.email, service: 'productDeleted', name: user.first_name, product } )
+        }
+        return response
+      } else return false
     } catch (error) {
       logger.error(error.message)
       throw new Error(error.message)
@@ -71,8 +105,18 @@ export const deleteByIdService = async (id) => {
 
 export const deleteAllService = async (email) => {
   try {
-    const response = await prodDao.deleteAllProducts(email);
-    return response;
+    const userDao = new UserDao();
+    const products = await prodDao.getAll()
+    if ( products ) {
+      const response = await prodDao.deleteAllProducts(email);
+      products.forEach( async prod => {
+        if ( prod.owner && prod.owner !== 'admin' ) {
+          const user = await userDao.getByEmail(prod.owner)
+          await sendMailEthereal( { destination: user.email, service: 'productDeleted', name: user.first_name, product: prod } )
+        }
+      });
+      return response;
+    } else return false
   } catch (error) {
     logger.error(error.message)
     throw new Error(error.message)
